// frontend/src/lib/api/generated/models/ListingEventCommentsItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { VoteTypeCode } from '../models/VoteTypeCode';
import { HttpFile } from '../http/http';

export class ListingEventCommentsItem {
    'id': number;
    'content': string;
    'createdAt': Date;
    'updatedAt': Date;
    'userId'?: number | null;
    'userName'?: string | null;
    'userAvatar'?: string | null;
    'replyCount'?: number | null;
    'voteCount'?: number | null;
    'isPinned': boolean;
    'voteType'?: VoteTypeCode | null;
    'deletedAt'?: Date | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "userId",
            "baseName": "user_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "userName",
            "baseName": "user_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "userAvatar",
            "baseName": "user_avatar",
            "type": "string",
            "format": ""
        },
        {
            "name": "replyCount",
            "baseName": "reply_count",
            "type": "number",
            "format": ""
        },
        {
            "name": "voteCount",
            "baseName": "vote_count",
            "type": "number",
            "format": ""
        },
        {
            "name": "isPinned",
            "baseName": "is_pinned",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "voteType",
            "baseName": "vote_type",
            "type": "VoteTypeCode",
            "format": ""
        },
        {
            "name": "deletedAt",
            "baseName": "deleted_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return ListingEventCommentsItem.attributeTypeMap;
    }

    public constructor() {
    }
}